unifying error handling and introducing logging
To unify error handling I moved the raise macro in the tools
namespace to access it easily from everywhere. It now is
capable to be used by clj and cljs. There is an error logged
everytime the raise macro is called.

For logging I introduced taoensso.timbre to be able to log in
Clojure and Clojurescript. This library supports both
structured and unstructured logging and suits our needs for
log analysis. The macros of this library can be called where
needed and can process all data structures.

The biggest problem is bug CLJ-865 in Clojure that prevents
passing line numbers between macros. Luckily timbre is
awesome and has a low level macro that makes it possible to
circumvent this bug. After a lot of fiddling it is now
possible to use a raise macro in the tools namespace and see
line numbers in the logs.

- Closes #54
- Great! https://github.com/ptaoussanis/timbre/commit/36785d68715bf9d7496c30eef5e3839c4e45e28e
- documentation of logging and error handling

Squashed commit of the following:

commit f59e83d3c8bd6cdd7ab23644353c7cf4a819c669
Author: Timo Kramer <fw-git@timokramer.de>
Date:   Thu Apr 30 17:32:48 2020 +0200

    set log-level for tests to fatal

commit 35f4cf299921ad49f8de90fef1aee0de3e1f600c
Author: Timo Kramer <fw-git@timokramer.de>
Date:   Thu Apr 30 17:02:59 2020 +0200

    using log directly is best

commit 4d484c6f4541d9cc4e8f5481474ce8eb5a8f7308
Author: Timo Kramer <fw-git@timokramer.de>
Date:   Thu Apr 30 15:55:48 2020 +0200

    transact has special sauce

commit abe0bac6546cf2955de0b7fbf0645843af067ed8
Author: Timo Kramer <fw-git@timokramer.de>
Date:   Wed Apr 29 15:18:15 2020 +0200

    implemented logging and error handling in tools namespace

commit 902411cd59292b683949b69fe49786db7256461f
Author: Timo Kramer <fw-git@timokramer.de>
Date:   Thu Apr 9 20:31:30 2020 +0200

    raise moved to tools namespace